Drop the _gen filename suffix
These files were automatically generated, but have been edited since.
