+++ object/host_network_system.go
